ct/reconciler: move into level_zero directory
