+++ src/v/cloud_topics/BUILD
@@ -10,8 +10,6 @@ package(default_visibility = [
     "//src/v/cloud_topics/level_zero:__subpackages__",
     "//src/v/cloud_topics/read_path:__pkg__",
     "//src/v/cloud_topics/read_path/tests:__pkg__",
-    "//src/v/cloud_topics/reconciler:__pkg__",
-    "//src/v/cloud_topics/reconciler/tests:__pkg__",
     "//src/v/cloud_topics/tests:__pkg__",
     "//src/v/cloud_topics/throttler:__pkg__",
     "//src/v/cloud_topics/throttler/tests:__pkg__",
@@ -126,7 +124,7 @@ redpanda_cc_library(
     include_prefix = "cloud_topics",
     visibility = [
         "//src/v/cloud_topics/frontend:__pkg__",
-        "//src/v/cloud_topics/reconciler:__pkg__",
+        "//src/v/cloud_topics/level_zero/reconciler:__pkg__",
         "//src/v/kafka/data:__pkg__",
         "//src/v/redpanda:__pkg__",
     ],
@@ -161,8 +159,8 @@ redpanda_cc_library(
         "//src/v/cloud_topics/batcher",
         "//src/v/cloud_topics/core:read_pipeline",
         "//src/v/cloud_topics/core:write_pipeline",
+        "//src/v/cloud_topics/level_zero/reconciler",
         "//src/v/cloud_topics/read_path:fetch_handler",
-        "//src/v/cloud_topics/reconciler",
         "//src/v/cloud_topics/throttler",
         "//src/v/model",
         "//src/v/storage",
