+++ src/v/cloud_topics/data_plane_impl.cc
@@ -17,8 +17,8 @@
 #include "cloud_topics/core/read_pipeline.h"
 #include "cloud_topics/core/write_pipeline.h"
 #include "cloud_topics/ephemeral_cluster_services.h"
+#include "cloud_topics/level_zero/reconciler/reconciler.h"
 #include "cloud_topics/read_path/fetch_request_handler.h"
-#include "cloud_topics/reconciler/reconciler.h"
 #include "cloud_topics/throttler/throttler.h"
 #include "model/fundamental.h"
 #include "model/record_batch_reader.h"
