+++ src/v/cloud_topics/level_zero/reconciler/BUILD
@@ -1,6 +1,6 @@
 load("//bazel:build.bzl", "redpanda_cc_library")
 
-package(default_visibility = ["//src/v/cloud_topics/reconciler/tests:__pkg__"])
+package(default_visibility = [":__subpackages__"])
 
 redpanda_cc_library(
     name = "range_batch_consumer",
@@ -10,7 +10,7 @@ redpanda_cc_library(
     hdrs = [
         "range_batch_consumer.h",
     ],
-    include_prefix = "cloud_topics/reconciler",
+    include_prefix = "cloud_topics/level_zero/reconciler",
     deps = [
         "//src/v/base",
         "//src/v/bytes:iobuf",
@@ -31,7 +31,7 @@ redpanda_cc_library(
     implementation_deps = [
         "//src/v/kafka/data:partition_proxy",
     ],
-    include_prefix = "cloud_topics/reconciler",
+    include_prefix = "cloud_topics/level_zero/reconciler",
     visibility = ["//visibility:public"],
     deps = [
         ":range_batch_consumer",
