+++ src/v/cloud_topics/level_zero/reconciler/range_batch_consumer.h
